# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Vault — Self-hosted file backup API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  /** @nullable */
  email?: string | null;
  /** @minLength 6 */
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
  /** @nullable */
  email?: string | null;
  isAdmin: boolean;
}

export interface FileItem {
  id: number;
  userId: number;
  name: string;
  path: string;
  size: number;
  /** @nullable */
  mimeType?: string | null;
  isFolder: boolean;
  parentPath: string;
  createdAt: string;
  updatedAt?: string;
}

export interface FolderInput {
  /** @minLength 1 */
  name: string;
  parentPath: string;
}

export interface UploadInput {
  /** Destination folder path */
  parentPath: string;
}

export interface StorageStats {
  totalFiles: number;
  totalFolders: number;
  totalSize: number;
  usedFormatted: string;
}

export interface User {
  id: number;
  username: string;
  /** @nullable */
  email?: string | null;
  isAdmin: boolean;
  storageUsed: number;
  createdAt: string;
}

export interface UserInput {
  /** @minLength 3 */
  username: string;
  /** @nullable */
  email?: string | null;
  /** @minLength 6 */
  password: string;
  isAdmin?: boolean;
}

export interface PasswordInput {
  /** @minLength 6 */
  newPassword: string;
}

export type ListFilesParams = {
/**
 * Directory path to list (defaults to root)
 */
path?: string;
};

export type GetRecentFilesParams = {
limit?: number;
};