# lib/api-zod/src/generated/types/folderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Vault — Self-hosted file backup API
 * OpenAPI spec version: 0.1.0
 */

export interface FolderInput {
  /** @minLength 1 */
  name: string;
  parentPath: string;
}